ssh-mcp-demo/requirements.txt
fastmcp
paramiko
